Remove namespace from gateway options
